requirements.txt
aiodns==3.0.0
aiofiles==23.1.0
aiohttp==3.8.4
aiohttp-socks==0.8.0
aiosignal==1.3.1
altair==5.0.1
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/altgraph-0.17.2-py2.py3-none-any.whl
antlr4-python3-runtime==4.9.3
anyio==3.7.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asttokens==2.0.8
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.6.2
bs4==0.0.1
cchardet==2.1.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
cssselect==1.2.0
cycler==0.11.0
dataclasses==0.6
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Django==4.2
django-admin==2.0.2
django-cors-headers==3.14.0
django-excel-response2==3.0.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-six==1.0.5
django-storages==1.13.2
djangorestframework==3.14.0
docopt==0.6.2
drf-yasg==1.21.7
drf-yasg2==1.19.4
editdistpy==0.1.3
elastic-transport==8.4.0
elasticsearch==8.8.0
entrypoints==0.4
excel-base==1.0.4
exceptiongroup==1.1.1
executing==1.1.0
fake-useragent==1.1.3
fastapi==0.95.2
fastjsonschema==2.16.2
ffmpy==0.3.0
filelock==3.12.0
files==1.1.1
Flask==2.3.2
Flask-Cors==4.0.0
fonttools==4.39.4
frozenlist==1.3.3
fsmpy==2.1.0
fsspec==2023.5.0
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/future-0.18.2-py3-none-any.whl
gensim==4.3.1
geographiclib==2.0
geopy==2.3.0
googletransx==2.4.2
gradio==3.30.0
gradio_client==0.2.5
h11==0.14.0
hazm==0.9.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.12.0
inexactsearch==1.0.2
inflection==0.5.1
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isoweek==1.3.3
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.1
jellyfish==0.11.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kenlm==0.1
kiwisolver==1.4.4
linkify-it-py==2.0.2
lxml==4.9.1
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
MechanicalSoup==1.2.0
mistune==0.8.4
mpmath==1.3.0
msgpack==1.0.5
msgpack-python==0.5.6
multidict==6.0.4
nbclient==0.6.8
nbconvert==5.6.1
nbformat==5.6.0
nest-asyncio==1.5.5
networkx==3.1
nltk==3.8.1
notebook==6.4.12
numpy==1.24.3
omegaconf==2.3.0
openapi-codec==1.3.2
opencv-python==4.7.0.72
orjson==3.8.14
outcome==1.2.0
packaging==21.3
pandas==2.0.1
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
pluggy==1.2.0
portalocker==2.7.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycares==4.3.0
pycparser==2.21
pydantic==1.10.8
pydub==0.25.1
pyee==8.2.2
pyfinite==1.9.1
pyfsm==0.4.2
Pygments==2.13.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==2.0.0
pyrsistent==0.18.1
PySocks==1.7.1
pyspellchecker==0.7.2
pytest==7.4.0
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-multipart==0.0.6
python-socks==2.3.0
pytube==15.0.0
pytz==2023.3
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.0
regex==2023.5.5
requests==2.31.0
requests-html==0.10.0
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
safetensors==0.3.1
schedule==1.2.0
scikit-learn==1.2.2
scipy==1.10.1
screen==1.0.1
selenium==4.10.0
semantic-version==2.10.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
silpa-common==0.3
simplejson==3.19.1
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/six-1.15.0-py2.py3-none-any.whl
smart-open==6.3.0
sniffio==1.3.0
snscrape @ git+https://github.com/JustAnotherArchivist/snscrape.git@0d824ab77334ed4ab6250e5e491171afeccfb298
sortedcontainers==2.4.0
soundex==1.1.3
soupsieve==2.3.2.post1
spellcheck==1.0.2
spellchecker==0.4
sqlparse==0.4.4
stack-data==0.5.1
starlette==0.27.0
sympy==1.12
symspellpy==6.7.7
termcolor==2.3.0
terminado==0.15.0
testpath==0.6.0
threadpoolctl==3.1.0
TimeConvert==3.0.13
timm==0.9.2
tinycss2==1.1.1
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
traitlets==5.4.0
transformers==4.29.2
trio==0.22.0
trio-websocket==0.10.3
twint @ git+https://github.com/twintproject/twint.git@e7c8a0c764f6879188e5c21e25fb6f1f856a7221
twitter-scraper==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==5.0.1
uc-micro-py==1.0.2
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.22.0
w3lib==2.1.1
wcwidth==0.2.5
webencodings==0.5.1
webifier==0.1.15
websockets==10.4
Werkzeug==2.3.6
widgetsnbextension==4.0.3
wsproto==1.2.0
xlwt==1.3.0
yarg==0.1.9
yarl==1.9.2
zipp==3.8.1
